requirements.txt
# Core Data Science & ML
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Medical Image Processing
SimpleITK>=2.2.0
nibabel>=5.0.0
pydicom>=2.3.0

# Computer Vision & Image Processing
opencv-python>=4.8.0
Pillow>=10.0.0

# Deep Learning (PyTorch - install via conda for CUDA support)
# torch>=2.0.0
# torchvision>=0.15.0

# Medical AI Framework
monai[all]>=1.3.0
torchmetrics>=1.0.0

# Utilities
matplotlib>=3.7.0
tqdm>=4.65.0
PyYAML>=6.0

# Note: Install PyTorch with CUDA via conda using environment.yml
# or manually: conda install pytorch torchvision pytorch-cuda=12.1 -c pytorch -c nvidia

